backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7
pydantic[email]==2.5.0
pydantic-settings==2.0.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
